requirements.txt
asgiref==3.10.0
Django==5.2.7
setuptools==80.9.0
sqlparse==0.5.3
Pillow==11.3.0
djangorestframework==3.16.1
django-filter==25.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
docutils==0.22.2
flake8==7.3.0
flake8-docstrings==1.7.0
django-debug-toolbar==6.0.0
sentry-sdk==2.41.0
poetry==2.2.1
django-cleanup==9.0.0